refactor(server): split modules
